migrate: move agent related feature into agentManager folder
