+++ wu-wei/src/agentManager/ExecutionRegistry.ts
