+++ wu-wei/src/agentManager/agentInterface.ts
